*add paginas de edição e listagem *add input hiddens do objeto usuario no _form.jsp *add REST pattern para objeto usuario. *update GenericDAO metodo load, parametro de int p/ Long - Motivo do ID *add novos metodos no usuario controller.

+++ src/main/java/br/com/fip/gati/revistaonline/repositorio/Repositorio.java
@@ -9,6 +9,6 @@ public interface Repositorio<T> {
 	public void save(T entity);
 	public void update(T entity);
 	public void delete(T entity);
-	public T load(int id);
+	public T load(Long id);
 	public List<T> listAll();
 }
